+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 2.5.1 (not yet released)
 
+- `make shar` to build a shar for upgrading the guest tools in one's
+  CentOS VM. Builds of these will be provided on
+  <https://download.joyent.com/pub/guest-tools>.
+
 - IMAGE-370 Update guest tools (/lib/smartdc/...) for image management.
   New the 'mdata-*' tools, adding mdata-list, mdata-put and mdata-delete.
   Also adds support for an 'operator-script' run at boot time.